+++ Source/ChilliSource/Input/Gesture/PinchGesture.cpp
@@ -161,7 +161,7 @@ namespace ChilliSource
                     }
                 }
                 
-                gesturePos /= k_requiredPointerCount;
+                gesturePos /= f32(k_requiredPointerCount);
             }
             
             return gesturePos;
